Modify Text property tests
When upgrading to ghc 9.0.*, we encountered test failures where some of
the expected successess fails, e.g., isSymbol '\129994' fails for GHC
9.0.* yet passed for GHC 8 and GHC 9.2. This is likely a bug in GHC/Base,
but it reveals a flaw in the testing strategy. The idea of our predicates
is that they should exactly match the ones defined in Data.Char. That
Data.Char is doing something wrong in this case is unfortunate, but it is
not our concern, and indeed we _want_ to match Data.Char's behavior here.

Thus we change the tests to check that our predicates match Data.Char, rather
than expecting them to math what *should* be happening according to the unicode
spec. This has the added bonus of allowing us to remove a _lot_ of generators
and test code.

We also split the generators into separate modules, for organization purposes.